Move Coordinates/CoordinateOffset types to coordinates module
